+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Unreleased
 
 - Add uninstall command for easily removing installed plugins
-- Behavior changed for `acp install --dev`. It will now also overwrite any versions previously installed without `--dev`.
+- Behavior changed for `acp install`. It will now overwrite existing plugins when switching to/from `--dev` mode.
 - More verbose command output
 
 # v1.0.0-beta.3