+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 - Add uninstall command for easily removing installed plugins
 - Add rename command for easily changing the publisher, the plugin name or the npm package name of the plugin
-- Behavior changed for `acp install`. It will now overwrite existing plugins when switching to/from `--dev` mode.
-- Behavior change for `acp init`. `npm_package_name` is now the last argument and is optional. `npm install` is automatically executed after generating the boilerplate.
 - More verbose command output
+- Better command options parsing and usage printing
 - Add default .gitignore and .npmignore files to generated boilerplate.
+
+## Breaking changes
+- (`acp init`): `npm_package_name` is now the last argument and is optional.
+- (`acp init`): `npm install` is automatically executed after generating the boilerplate.
+- (`acp install`): will now overwrite existing plugin installations when switching to/from `--dev` mode.
 
 # v1.0.0-beta.3